remove oslo-incubator code
The bulk of the oslo-incubator code was related to logging, so
I added oslo.log as a requirement and updated the necessary
spots. The other part that is more complex was using oslo.config
to generate a sample config file. To do that, I had to create the
necessary config-generator .conf file, this also meant that the
items under the tools directory could be removed.

Change-Id: I424e07f6a4d688c7686d2fab674fe7f2a475ed00